+++ GoalMogul-iOS/src/Main/Common/ActionSheetFactory.js
@@ -1,0 +1,3 @@
+import {
+  ActionSheetIOS
+} from 'react-native';