+++ GoalMogul-iOS/src/Main/Common/SearchBarHeader.js
@@ -4,10 +4,10 @@ import {
   Image,
   TouchableWithoutFeedback,
   Text,
-  TouchableOpacity
+  TouchableOpacity,
+  ActionSheetIOS
 } from 'react-native';
 import { Icon } from 'react-native-elements';
-import { BlurView } from 'react-native-blur';
 import { connect } from 'react-redux';
 import { Actions } from 'react-native-router-flux';
 
@@ -15,12 +15,26 @@ import { Actions } from 'react-native-router-flux';
 import Logo from '../../asset/header/logo.png';
 import IconMenu from '../../asset/header/menu.png';
 import Setting from '../../asset/header/setting.png';
-import BackButton from '../../asset/utils/back.png'
+import BackButton from '../../asset/utils/back.png';
+import FriendsSettingIcon from '../../asset/utils/friendsSettingIcon.png';
 
 /* Actions */
-import { back, openProfile, openSetting } from '../../actions';
+import {
+  back,
+  openProfile,
+  openSetting,
+  blockUser
+} from '../../actions';
 
 const tintColor = '#33485e';
+
+// For profile friend setting ActionSheetIOS
+const FRIENDSHIP_SETTING_BUTTONS = ['Block', 'Report', 'Cancel'];
+const BLOCK_INDEX = 0;
+const REPORT_INDEX = 1;
+const CANCEL_INDEX = 2;
+
+const DEBUG_KEY = '[ Component SearchBarHeader ]';
 
 /**
   TODO: refactor element to have consistent behavior
@@ -47,6 +61,34 @@ class SearchBarHeader extends Component {
   handleSettingOnClick() {
     // TODO: open account setting page
     this.props.openSetting();
+  }
+
+  handleFriendsSettingOnClick = () => {
+    ActionSheetIOS.showActionSheetWithOptions({
+      options: FRIENDSHIP_SETTING_BUTTONS,
+      cancelButtonIndex: CANCEL_INDEX,
+    },
+    (buttonIndex) => {
+      console.log('button clicked', FRIENDSHIP_SETTING_BUTTONS[buttonIndex]);
+      switch (buttonIndex) {
+        case BLOCK_INDEX:
+          // User chose to block user with id: _id
+          console.log(`${DEBUG_KEY} User blocks _id: `, this.props.profileUserId);
+          this.props.blockUser(this.props.profileUserId);
+          break;
+
+        case REPORT_INDEX:
+          // User chose to unfriend
+          console.log(`${DEBUG_KEY} User reports profile with _id: `, this.props.profileUserId);
+          // this.props.updateFriendship(_id, 'deleteFriend', TAB_KEY, () => {
+          //   console.log('Successfully delete friend with id: ', _id);
+          //   this.setState({ requested: false });
+          // });
+          break;
+        default:
+          return;
+      }
+    });
   }
 
   renderSearchBarLeftIcon() {
@@ -81,6 +123,20 @@ class SearchBarHeader extends Component {
   }
 
   renderSearchBarRightIcon() {
+    // On other people's profile page
+    // if (this.props.setting && !this.props.haveSetting) {
+    if (this.props.setting && true) {
+      return (
+        <TouchableWithoutFeedback onPress={this.handleFriendsSettingOnClick.bind(this)}>
+          <Image
+            style={{ ...styles.headerRightImage, tintColor, height: 28 }}
+            source={FriendsSettingIcon}
+          />
+        </TouchableWithoutFeedback>
+      );
+    }
+
+    // On self profile page
     if (this.props.setting && this.props.haveSetting) {
       return (
         <TouchableWithoutFeedback onPress={this.handleSettingOnClick.bind(this)}>
@@ -88,11 +144,15 @@ class SearchBarHeader extends Component {
         </TouchableWithoutFeedback>
       );
     }
+
+    // Standard search bar menu
     if (this.props.rightIcon === 'menu') {
       return (
         <Image style={styles.headerRightImage} source={IconMenu} />
       );
     }
+
+    // Empty dummy view as default
     return (
       <View style={styles.headerRightImage} />
     );
@@ -198,12 +258,22 @@ const styles = {
 
 const mapStateToProps = state => {
   const { userId } = state.user;
+  const profileUserId = state.profile.userId;
   const haveSetting = state.profile.userId.toString() === state.user.userId.toString();
 
   return {
     userId,
-    haveSetting
+    haveSetting,
+    profileUserId
   };
 };
 
-export default connect(mapStateToProps, { back, openProfile, openSetting })(SearchBarHeader);
+export default connect(
+  mapStateToProps,
+  {
+    back,
+    openProfile,
+    openSetting,
+    blockUser
+  }
+)(SearchBarHeader);