+++ GoalMogul-iOS/CHANGELOG.md
@@ -10,8 +10,14 @@ if there is corresponding issue, state it
 
 * Version 0.0.1
 ## 1 June 2018, by Jia Zeng <andyzeng96@gmail.com>
+* Update SearchBarHeader to include Profile Friendship setting
+  that includes block and report user
 * Bug fix for rendering # of mutual friends
 * Update profile detail design
+  * Occupation
+  * divider
+  * Elevator pitch
+  * about
 
 ## 1 June 2018, by Jia Zeng <andyzeng96@gmail.com>
 * Empty results for suggested (“No Recommendations ”)